Application made more reusable
BeforeTheGameFragment was refactored to be reusable. It is now
MenuFragment, which will be usede everywhere. More data added
and a play button has been overlayed on the video. The videoview
was also optimized so that when the user presses on it, the video
will play, but if the user is scrolling and accidentally touches
it doesn't count and won't play.